Fix cancellation spec
The visitor's confirmed booking page has been changed to include a new
element which hid the "yes I want to cancel this visit" checkbox.  As a
result the related integration test started failing as it could no
longer scroll that element into view.  I have added the required step to
click this new link in order for the checkbox to be found.