Update tests re: Mailtrap API changes
Mailtrap changed their API over the weekend and this resulted in two of
our tests failing.  The "text_body" field is now returned with a
deprecation warning and a secondary API call has to be made to get the
body of the email.

There is the opportunity for further refactoring in the mailtrap.rb file
and the way we are making the now required second API call to get the
body of the email.

In addition to this it was established that a method was no longer being
used (search_body) so this has been removed.